Разделит 'VideoPlayer' на 'Player' и 'withVideo'.
